Benchmarking: Added runtime and mem parameters to test_func_data
Gave test_func_data the capability to not do one of the benchmarks by
setting the parameter to False. However, actually using this will require
some changes in gen_graphs. If anyone wants to do this, they'll have to
change gen_graphs.py, but I just implemented it in test_func_data
because it was simple.